requirements.txt
allure-behave==2.13.5
allure-pytest==2.13.5
behave~=1.2.6
fastapi~=0.110.0
hvac~=2.1.0
nose2==0.14.1
passlib~=1.7.4
pydantic~=2.6.4
pymsteams~=0.2.2
pytest==8.0.2
python-dotenv==1.0.1
requests==2.31.0
